fix point type and category name

+++ app/Models/Results/Models/BrojureResultsModel.php
@@ -184,7 +184,7 @@ left join submit s ON s.task_id = t.task_id AND s.contestant_id = ct.contestant_
      *
      * @return int sum of Student Pilny points
      */
-    private function getSumLimitForStudentPilny(): int
+    private function getSumLimitForStudentPilny(): float
     {
         return $this->getSumLimit(ModelCategory::tryFrom(ModelCategory::FYKOS_4));
     }
@@ -193,7 +193,7 @@ left join submit s ON s.task_id = t.task_id AND s.contestant_id = ct.contestant_
      * Returns total points for given category and series
      * @return int sum of points
      */
-    private function getSumLimit(ModelCategory $category): int
+    private function getSumLimit(ModelCategory $category): float
     {
         $sum = 0;
         foreach ($this->getSeries() as $series) {
